django_rest_app/requirements.txt
Django==1.9.6
ipython==4.2.0
ipdb==0.10.0
requests==2.10.0
djangorestframework==3.3.3

